requirements.txt
streamlit==1.47.1
google-generativeai==0.8.5
python-dotenv==1.1.1
pdf2image==1.17.0
pillow==10.4.0
pytest==8.4.1